requirements.txt
flask==2.3.2
ultralytics==8.0.134
torch==2.0.1+cpu
torchvision==0.15.2+cpu
numpy==1.24.4
pillow==10.0.0
opencv-python-headless==4.8.0.74